requirements.txt
torch>=1.9.0
torch-geometric>=2.0.0
numpy>=1.19.2
scikit-learn>=0.24.2
pyyaml>=5.4.1
tqdm>=4.62.3
matplotlib>=3.4.3
pandas>=1.2.0